# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "description": "API for Social Media application",
        "title": "Social Media Application API"
    },
    "host": "localhost:3000",
    "securityDefinitions": {
        "jwt": {}
    },
    "paths": {
...
        "/api/posts/": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Logged-in user posts.",
                "description": "All posts of currently logged in user.",
                "operationId": "getPostsByUser",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "All posts in the result key of response object."
                    },
                    "404": {
                        "description": "Posts not found."
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Draft a new Post.",
                "description": "New post will create.",
                "operationId": "createPost",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "imageUrl",
                        "in": "formData",
                        "type": "file",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "caption",
                        "in": "formData",
                        "type": "string",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "In result key of response object you will find newly created post and message as post drafted successfully."
                    },
                    "400": {
                        "description": "Error while creating post."
                    }
                }
            }
        },
        "/api/posts/all": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Fetch all posts",
                "description": "Fetching all posts and display according to pagination parameters.",
                "operationId": "getAllPosts",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all post."
                    },
                    "404": {
                        "description": "Error while fetching posts."
                    }
                }
            }
        },
        "/api/posts/{postId}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "To fetch a single Post.",
                "description": "Able to fetch mentioned id post ",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find fetched post."
                    },
                    "404": {
                        "description": "Post not found. Please verify postId you want."
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Update a post.",
                "description": "Only authenticated user updates a mentioned id post.",
                "operationId": "updatePostById",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to update.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "imageUrl",
                        "in": "formData",
                        "type": "file",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "caption",
                        "in": "formData",
                        "type": "string",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post updated successfully."
                    },
                    "400": {
                        "description": "Invalid postId supplied."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete a post.",
                "description": "Only authenticated user deletes a mentioned id post.",
                "operationId": "deletePostById",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to delete.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted successfully."
                    },
                    "400": {
                        "description": "Invalid postId supplied."
                    }
                }
            }
        },
        "/api/posts/filter": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Filter posts by caption.",
                "description": "To get all post matching with mentioned captionand display according to pagination parameters..",
                "operationId": "getAllPostsByCaption",
                "parameters": [
                    {
                        "name": "caption",
                        "in": "query",
                        "type": "string",
                        "description": "Caption of post to filter",
                        "required": true
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "type": "integer",
                        "description": "",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "In result key of response object you will find all filtered post."
                    },
                    "404": {
                        "description": "Error while fetching posts."
                    }
                }
            }
        },
        "/api/posts/archieve/{postId}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Archieve a post.",
                "description": "Only authenticated user archieve a mentioned id post.",
                "operationId": "archievePost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to archieve.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post archieved successfully."
                    },
                    "400": {
                        "description": "Invalid postId supplied."
                    }
                }
            }
        },
        "/api/posts/publish/{postId}": {
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Publish a Post",
                "description": "Only authenticated user publish a mentioned id post.",
                "operationId": "publishPost",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "description": "The id of the post to publish.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post published successfully."
                    },
                    "400": {
                        "description": "Invalid postId supplied."
                    }
                }
            }
        },
...
    },
...
}